Refactor useForm logic into the useForm component
